Port database layer to SQLite/D1 (phases 2+3)
Schema: pg-core -> sqlite-core across all 60 tables. pgEnums become
text({ enum }) columns; serial -> integer autoincrement PKs; timestamps
-> integer timestamp_ms (unixepoch subsec default); json/boolean columns
-> sqlite modes. Fresh migration journal (0000_init) — D1 is a new
database, applied via wrangler d1 migrations, not drizzle's migrator.

Driver: drizzle-orm/d1 on Workers (per-request instance from the DB
binding via runWithRequestDb), drizzle-orm/libsql against a file: path
for bun contexts (tests/scripts) — async + batch(), the closest local
match to D1 semantics. postgres/pg dependencies removed.

Transactions (D1 has none interactive) — all 6 sites restructured:
- rius rotate(): atomic db.batch of the status walk + seed insert
- sius/bius startRound + topUpActiveRounds: count guard and insert
  fused into one atomic INSERT ... SELECT ... WHERE statement
- chain-game createFirstSet/continueSet: optimistic read-check-insert
  backed by new partial unique indexes (one live child per set, one
  live set per round position); lost races surface as UNIQUE
  violations translated back to the game's invariant copy
- advisory locks removed everywhere

Single-process state ports (Workers isolates share no memory):
- presence: in-memory maps -> presence table (hashed-IP anon subjects,
  lazy pruning)
- auth rate limiting: in-memory buckets -> rate_limits table with an
  atomic guarded upsert. Global, so isolate/colo distribution no longer
  multiplies an attacker's budget. (Planned per-code attempts column
  proved unsound — enterCode receives only the code, so misses match no
  row; the keyed-counter table replaces it.)
- tourney realtime: in-memory EventEmitter pubsub -> SSE route polls
  the tournaments row (updatedAt + new adminHeartbeatAt) every 2s;
  persisting is the publish

Raw SQL sweep: ilike->like, jsonb ?|/@> -> json_each EXISTS predicates,
array_agg -> json_group_array (+ JSON.parse), DATE_TRUNC/TO_CHAR ->
strftime, FILTER -> SUM(CASE), lateral json_agg -> correlated subquery,
greatest -> max, ::casts dropped.

Tests: integration runner drops docker postgres for an ephemeral sqlite
file (suite: ~3min -> ~12s); truncate helper rewritten (FKs off during
wipe); constraint assertions match SQLite's column-based messages.
Local dev: bun dev now applies D1 migrations locally and runs vite dev
(workerd serves SSR against local D1); the sandbox/tunnel postgres dev
flows and migrate-prod are deleted. DEPLOY.md rewrite lands with the CI
cutover.

+++ apps/web/src/lib/notifications/helpers.server.ts
@@ -191,7 +191,7 @@ export async function deleteNotificationsForMessage(
       and(
         eq(notifications.type, "message_like"),
         eq(notifications.entityType, entityType),
-        sql`(${notifications.data}->>'messageId')::int = ${messageId}`,
+        sql`${notifications.data}->>'messageId' = ${messageId}`,
       ),
     )
 }
